fixed Location and edit ITem

+++ app/src/main/java/com/amt/doineedit_firebase/appDB/Item.kt
@@ -4,7 +4,7 @@ import com.google.firebase.database.Exclude
 import com.google.firebase.database.IgnoreExtraProperties
 
 @IgnoreExtraProperties
-data class Item(val itemName: String="", val price:Float= 0.0F, val quantity:Int=1, val haveItem:Boolean = false){
+data class Item(val itemName: String="", val price:Double= 0.0, val quantity:Int=1, val haveItem:Boolean = false){
     var showMenu = false
     @Exclude
     fun toMap(): Map<String, Any?> {
